editor/libeditor/text/tests/test_bug641466.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=641466
-->
<head>
  <title>Test for Bug 641466</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=641466">Mozilla Bug 641466</a>
<p id="display"></p>
<div id="content">
<input value="&#x10451;&#x10467;&#x10455;&#x10451;">
<textarea>&#x10451;&#x10467;&#x10455;&#x10451;</textarea>
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 641466 **/
SimpleTest.waitForExplicitFinish();
SimpleTest.waitForFocus(function() {
  function doTest(element) {
    element.focus();
    element.selectionStart = 4;
    element.selectionEnd = 4;
    synthesizeKey("VK_BACKSPACE", {});
    synthesizeKey("VK_BACKSPACE", {});
    synthesizeKey("VK_BACKSPACE", {});
    synthesizeKey("VK_BACKSPACE", {});

    ok(element.value, "", "4 backspaces should delete all of the characters in the " + element.localName);
  }

  doTest(document.querySelector("input"));
  doTest(document.querySelector("textarea"));

  SimpleTest.finish();
});

</script>
</pre>
</body>
</html>
